+++ packages/sandbox/src/core/compartment/index.ts
@@ -17,7 +17,7 @@ declare global {
   interface Window {
     __compartment_window__?: Window;
 
-    [p: CompartmentGlobalId]: WindowProxy;
+    [p: CompartmentGlobalId]: WindowProxy | undefined;
   }
 }
 
@@ -27,13 +27,7 @@ export class Compartment {
   /**
    * Since the time of execution of the code in Compartment is determined by the browser, a unique compartmentSpecifier should be generated in Compartment
    */
-  private readonly id: CompartmentGlobalId = (() => {
-    // make sure the compartmentSpecifier is unique
-    while (nativeGlobal[getCompartmentGlobalId(compartmentCounter)]) {
-      compartmentCounter++;
-    }
-    return getCompartmentGlobalId(compartmentCounter);
-  })();
+  private readonly id: CompartmentGlobalId;
 
   private readonly _globalThis: WindowProxy;
 
@@ -41,6 +35,13 @@ export class Compartment {
 
   constructor(globalProxy: WindowProxy) {
     this._globalThis = globalProxy;
+
+    // make sure the compartmentSpecifier is unique
+    while (nativeGlobal[getCompartmentGlobalId(compartmentCounter)]) {
+      compartmentCounter++;
+    }
+    this.id = getCompartmentGlobalId(compartmentCounter);
+    nativeGlobal[this.id] = globalProxy;
   }
 
   get globalThis(): WindowProxy {
@@ -48,15 +49,16 @@ export class Compartment {
   }
 
   protected addConstantIntrinsicNames(intrinsics: string[]): void {
-    this.constantIntrinsicNames  = [...intrinsics, ...this.constantIntrinsicNames];
+    this.constantIntrinsicNames = [...intrinsics, ...this.constantIntrinsicNames];
   }
 
   makeEvaluateFactory(source: string, sourceURL?: string): string {
     const sourceMapURL = sourceURL ? `//# sourceURL=${sourceURL}\n` : '';
 
-    const globalObjectOptimizer = this.constantIntrinsicNames.length ? `const {${this.constantIntrinsicNames.join(',')}} = this;`: '';
+    const globalObjectOptimizer = this.constantIntrinsicNames.length
+      ? `const {${this.constantIntrinsicNames.join(',')}} = this;`
+      : '';
 
-    nativeGlobal[this.id] = this.globalThis;
     // eslint-disable-next-line max-len
     return `;(function(){with(this){${globalObjectOptimizer}${source}\n${sourceMapURL}}}).bind(window.${this.id})();`;
   }