+++ packages/sandbox/src/utils.ts
@@ -16,30 +16,21 @@ export function isConstructable(fn: CallableFunction): fn is CallableFunction {
 
   if (hasPrototypeMethods) return true;
 
-  const cachedValue = fnRegexCheckCacheMap.get(fn);
-  if (typeof cachedValue !== 'undefined') {
-    return cachedValue;
+  const cachedResult = fnRegexCheckCacheMap.get(fn);
+  if (typeof cachedResult !== 'undefined') {
+    return cachedResult;
   }
 
-  let constructable: boolean = hasPrototypeMethods;
-  if (!constructable) {
-    // fn.toString has a significant performance overhead, if hasPrototypeMethods check not passed, we will check the function string with regex
-    const fnString = fn.toString();
-    const constructableFunctionRegex = /^function\b\s[A-Z].*/;
-    const classRegex = /^class\b/;
-    constructable = constructableFunctionRegex.test(fnString) || classRegex.test(fnString);
-  }
+  // fn.toString has a significant performance overhead, if hasPrototypeMethods check not passed, we will check the function string with regex
+  const fnString = fn.toString();
+  const constructableFunctionRegex = /^function\b\s[A-Z].*/;
+  const classRegex = /^class\b/;
+  const constructable = constructableFunctionRegex.test(fnString) || classRegex.test(fnString);
 
   fnRegexCheckCacheMap.set(fn, constructable);
   return constructable;
 }
 
-/**
- * in safari
- * typeof document.all === 'undefined' // true
- * typeof document.all === 'function' // true
- * We need to discriminate safari for better performance
- */
 const callableFnCacheMap = new WeakMap<CallableFunction, boolean>();
 
 export function isCallable(fn: unknown): fn is CallableFunction {
@@ -47,6 +38,11 @@ export function isCallable(fn: unknown): fn is CallableFunction {
     return true;
   }
 
+  /*
+   * We can not use typeof to confirm it is function as in some safari version
+   * typeof document.all === 'undefined' // true
+   * typeof document.all === 'function' // true
+   */
   const callable = typeof fn === 'function' && fn instanceof Function;
   if (callable) {
     callableFnCacheMap.set(fn, callable);
@@ -57,7 +53,7 @@ export function isCallable(fn: unknown): fn is CallableFunction {
 const frozenPropertyCacheMap = new WeakMap<object, Record<PropertyKey, boolean>>();
 
 export function isPropertyFrozen(target: object, p?: PropertyKey): boolean {
-  if (!target || !p) {
+  if (!p) {
     return false;
   }