+++ packages/shared/src/assets-transpilers/types.ts
@@ -2,21 +2,21 @@
  * @author Kuitos
  * @since 2023-08-26
  */
-import type { Sandbox } from '@qiankunjs/sandbox';
 import type { BaseLoaderOpts } from '../common';
-
 import type { MatchResult } from '../module-resolver';
 
 export type BaseTranspilerOpts = BaseLoaderOpts & {
   moduleResolver?: (url: string) => MatchResult | undefined;
-  sandbox?: Sandbox;
+  sandbox?: {
+    makeEvaluateFactory(source: string, sourceURL?: string): string;
+  };
 };
 
 export type AssetsTranspilerOpts = BaseTranspilerOpts & { rawNode: Node };
 
 export enum Mode {
-  REMOTE_FROM_SANDBOX = 'REMOTE_FROM_SANDBOX',
-  CACHE_FROM_SANDBOX = 'CACHE_IN_SANDBOX',
-  INLINE_FROM_SANDBOX = 'INLINE_FROM_SANDBOX',
+  REMOTE_ASSETS_IN_SANDBOX = 'RAIS',
+  REUSED_DEP_IN_SANDBOX = 'RDIS',
+  INLINE_CODE_IN_SANDBOX = 'ICIS',
   NONE = 'NONE',
 }