+++ packages/loader/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @qiankunjs/loader
 
+## 0.0.1-rc.4
+
+### Patch Changes
+
+- Updated dependencies [39301f19]
+  - @qiankunjs/sandbox@0.0.1-rc.3
+  - @qiankunjs/shared@0.0.1-rc.3
+
+## 0.0.1-rc.3
+
+### Patch Changes
+
+- Updated dependencies [b23d3d7b]
+  - @qiankunjs/shared@0.0.1-rc.2
+  - @qiankunjs/sandbox@0.0.1-rc.2
+
+## 0.0.1-rc.2
+
+### Patch Changes
+
+- Updated dependencies [ebb2bcaa]
+  - @qiankunjs/shared@0.0.1-rc.1
+  - @qiankunjs/sandbox@0.0.1-rc.1
+
 ## 0.0.1-beta.6
 
 ### Patch Changes