+++ packages/sandbox/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @qiankunjs/sandbox
 
+## 0.0.1-rc.3
+
+### Patch Changes
+
+- 39301f19: 🔀 merge master
+  - @qiankunjs/shared@0.0.1-rc.3
+
+## 0.0.1-rc.2
+
+### Patch Changes
+
+- Updated dependencies [b23d3d7b]
+  - @qiankunjs/shared@0.0.1-rc.2
+
+## 0.0.1-rc.1
+
+### Patch Changes
+
+- Updated dependencies [ebb2bcaa]
+  - @qiankunjs/shared@0.0.1-rc.1
+
 ## 0.0.1-beta.6
 
 ### Patch Changes